+++ postgres_metrics/__init__.py
@@ -1,2 +1,2 @@
 __author__ = 'Markus Holtermann'
-__version__ = '0.5.0'
+__version__ = '0.6.0'