+++ docs/source/changelog.rst
@@ -46,4 +46,4 @@ Changelog
 0.1.0
 =====
 
-* Inital Release
+* Initial Release