+++ docs/source/changelog.rst
@@ -16,6 +16,13 @@ Changelog
 
 * The documentation now has an intersphinx setup to Python 3 and Django
 
+* The hard dependency on psycopg2 was dropped because installing wheel files
+  of psycopg2 can be troublesome `as outlined by the project maintainers
+  <http://initd.org/psycopg/articles/2018/02/08/psycopg-274-released/>`__. When
+  using django-postgres-metrics you now need to install psycopg2 or
+  psycopg2-binary explicitly. This is usually not an issue as either one is
+  required by Django anyway to use Django's PostgreSQL database backend.
+
 0.5.0
 =====